Package names to lowercase to avoid confusion Moved ViewModelProviderFactory.kt to utils.factory Safer way to access NotesViewModel.kt using viewModels/activityViewModels delegates Added a simpler and generic viewModelFactory
