+++ app/src/main/java/thecodemonks/org/nottzapp/ui/add/AddNotesFragment.kt
@@ -1,24 +1,21 @@
-package thecodemonks.org.nottzapp.UI.ui.AddNotes
+package thecodemonks.org.nottzapp.ui.add
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
 import kotlinx.android.synthetic.main.add_notes_fragment.*
 import thecodemonks.org.nottzapp.R
-import thecodemonks.org.nottzapp.UI.ui.Notes.ViewModels.NotesViewModel
-import thecodemonks.org.nottzapp.app.MainActivity
+import thecodemonks.org.nottzapp.ui.notes.NotesViewModel
 import thecodemonks.org.nottzapp.utils.toast
 
 class AddNotesFragment : Fragment(R.layout.add_notes_fragment) {
 
-    private lateinit var viewModel: NotesViewModel
+    private val viewModel: NotesViewModel by activityViewModels()
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        // init viewModel
-        viewModel = (activity as MainActivity).viewModel
-
 
         // save notes to db
         btn_save_notes.setOnClickListener {
@@ -42,6 +39,5 @@ class AddNotesFragment : Fragment(R.layout.add_notes_fragment) {
                 }
             }
         }
-
     }
 }
