+++ app/src/main/java/thecodemonks/org/nottzapp/ui/notes/NotesFragment.kt
@@ -1,8 +1,9 @@
-package thecodemonks.org.nottzapp.UI.ui.Notes
+package thecodemonks.org.nottzapp.ui.notes
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.ItemTouchHelper
@@ -11,19 +12,16 @@ import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.snackbar.Snackbar
 import kotlinx.android.synthetic.main.notes_fragment.*
 import thecodemonks.org.nottzapp.R
-import thecodemonks.org.nottzapp.UI.ui.Notes.ViewModels.NotesViewModel
 import thecodemonks.org.nottzapp.adapter.NotesAdapter
 import thecodemonks.org.nottzapp.app.MainActivity
 
 class NotesFragment : Fragment(R.layout.notes_fragment) {
 
-    private lateinit var viewModel: NotesViewModel
+    private val viewModel: NotesViewModel by activityViewModels()
     private lateinit var notesAdapter: NotesAdapter
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        viewModel = (activity as MainActivity).viewModel
 
         setUpRV()
 
@@ -83,8 +81,6 @@ class NotesFragment : Fragment(R.layout.notes_fragment) {
         ItemTouchHelper(itemTouchHelperCallback).apply {
             attachToRecyclerView(notes_rv)
         }
-
-
     }
 
     private fun setUpRV() {
