+++ app/src/main/res/layout/add_notes_fragment.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     android:background="@color/colorPrimary"
-    tools:context=".UI.ui.AddNotes.AddNotesFragment">
+    tools:context=".ui.add.AddNotesFragment">
 
 
     <EditText
